20240525 update

+++ inc/c/architect/arcriscv_64.h
@@ -27,8 +27,9 @@ typedef signed short int16_t;
 typedef unsigned short uint16_t;
 typedef signed int int32_t;
 typedef unsigned int uint32_t;
-typedef signed long int64_t;
-typedef unsigned long uint64_t;
+// long is 64-bit on Linux-GCC-Devk and RISCV64
+typedef signed long long int64_t;
+typedef unsigned long long uint64_t;
 
 
 //{TODO} more with reference: arcx64_win64.h
